Separate API code out into dedicated app
